feat: move user-related logic to users.js
